separated dagger modules
