+++ app/src/main/java/com/kondenko/yamblzweather/data/location/RoomLocationProvider.java
@@ -5,6 +5,8 @@ import com.kondenko.yamblzweather.domain.guards.LocationProvider;
 
 import java.util.List;
 
+import javax.inject.Inject;
+
 import io.reactivex.Completable;
 import io.reactivex.Observable;
 import io.reactivex.Single;
@@ -20,7 +22,8 @@ public class RoomLocationProvider implements LocationProvider {
     private final CityDao cityDao;
     private final BehaviorProcessor<CityEntity> cityBehaviour;
 
-    public RoomLocationProvider(CityDao cityDao) {
+    @Inject
+    RoomLocationProvider(CityDao cityDao) {
         this.cityDao = cityDao;
         cityBehaviour = BehaviorProcessor.create();
         cityDao.getSelectedCity().subscribe(cityBehaviour);
