+++ app/src/main/java/com/kondenko/yamblzweather/infrastructure/WeatherJobsScheduler.java
@@ -1,14 +1,15 @@
 package com.kondenko.yamblzweather.infrastructure;
 
-import com.evernote.android.job.JobCreator;
 import com.evernote.android.job.JobManager;
 import com.kondenko.yamblzweather.domain.guards.JobsScheduler;
 
 import java.util.concurrent.TimeUnit;
 
-public class WeatherJobsScheduler implements JobsScheduler {
+import javax.inject.Inject;
 
-    public WeatherJobsScheduler(JobCreator appJobCreator) {
+public class WeatherJobsScheduler implements JobsScheduler {
+    @Inject
+    WeatherJobsScheduler(AppJobCreator appJobCreator) {
         JobManager.instance().addJobCreator(appJobCreator);
     }
